+++ app/test/frontend/handlers/custom_api_test.dart
@@ -30,16 +30,8 @@ void main() {
                 'pubspec': {
                   'name': 'oxygen',
                   'version': '1.2.0',
-                  'description': 'oxygen is awesome',
-                  'homepage': 'https://oxygen.example.dev/',
                   'environment': {'sdk': '>=2.6.0 <3.0.0'},
                   'dependencies': {},
-                  'screenshots': [
-                    {
-                      'path': 'static.webp',
-                      'description': 'This is an awesome screenshot'
-                    }
-                  ]
                 },
                 'archive_url':
                     '${activeConfiguration.primaryApiUri}/packages/oxygen/versions/1.2.0.tar.gz',
@@ -56,18 +48,10 @@ void main() {
                 'pubspec': {
                   'name': 'flutter_titanium',
                   'version': '1.10.0',
-                  'description': 'flutter_titanium is awesome',
-                  'homepage': 'https://flutter_titanium.example.dev/',
                   'environment': {'sdk': '>=2.6.0 <3.0.0'},
                   'dependencies': {
                     'flutter': {'sdk': 'flutter'}
                   },
-                  'screenshots': [
-                    {
-                      'path': 'static.webp',
-                      'description': 'This is an awesome screenshot'
-                    }
-                  ]
                 },
                 'archive_url':
                     '${activeConfiguration.primaryApiUri}/packages/flutter_titanium/versions/1.10.0.tar.gz',
@@ -84,16 +68,8 @@ void main() {
                 'pubspec': {
                   'name': 'neon',
                   'version': '1.0.0',
-                  'description': 'neon is awesome',
-                  'homepage': 'https://neon.example.dev/',
                   'environment': {'sdk': '>=2.6.0 <3.0.0'},
                   'dependencies': {},
-                  'screenshots': [
-                    {
-                      'path': 'static.webp',
-                      'description': 'This is an awesome screenshot'
-                    }
-                  ]
                 },
                 'archive_url':
                     '${activeConfiguration.primaryApiUri}/packages/neon/versions/1.0.0.tar.gz',